GUI has been added!
I finally added a GUI.  I was planning on implementing it through JavaFX, but I was unhappy with the control I was given.

With the decision to implement a GUI, I'll be converting over all my functions to ones what play properly with the GUI.

I also did:
Implemented container creation in all the ports; which solved a null pointer exception
Made two random variables in the gamemap class private.
Cleaned up a tiny bit of the display data for the port class, I'll definitely be doing more of this in the future.