Additional GUI modifications
Spent most of the morning working on how the GUI works, I think I have it set up the way I want it to be.

Still working through porting over the application from being a terminal app to being an actual application.  I'm running into an issue with pausing, and waiting for user input.  I think I found a way to wait for user input, with an action performed call, but it's not working at the moment.